[Fix] OU 프로토타입 완성

+++ apps/gocho_web/src/component/card/jobCard/index.tsx
@@ -3,8 +3,9 @@ import Image from "next/image";
 import Link from "next/link";
 import { BsFillBookmarkFill } from "react-icons/bs";
 import { FiEye } from "react-icons/fi";
+import { useQueryClient } from "@tanstack/react-query";
 
-import { useAddUserBookmark, useDeleteUserBookmark } from "shared-api/bookmark";
+import { useAddJobUserBookmark, useDeleteUserBookmark } from "shared-api/bookmark";
 import { DdayBox } from "shared-ui/common/atom/dDayBox";
 
 import defaultCompanyLogo from "@public/images/global/common/default_company_logo.svg";
@@ -47,11 +48,18 @@ export const JobCard: FunctionComponent<JobCardProps | JobCardSkeleton> = ({
   isBookmarked,
   userId,
   isSkeleton,
-  refetchUserBookmark,
+  // refetchUserBookmark,
 }) => {
   const [imageSrc, setImageSrc] = useState(jobData?.companyLogo as string);
 
-  const { mutate: addMutate } = useAddUserBookmark();
+  const queryClient = useQueryClient();
+
+  const { mutate: addMutate } = useAddJobUserBookmark({
+    id: jobData?.id as number,
+    title: jobData?.title as string,
+    end_time: jobData?.endTime as number,
+    company: { id: jobData?.companyId as number, name: jobData?.companyName as string },
+  });
   const { mutate: deleteMutate } = useDeleteUserBookmark();
 
   if (isSkeleton || jobData === undefined) {
@@ -63,31 +71,19 @@ export const JobCard: FunctionComponent<JobCardProps | JobCardSkeleton> = ({
   }
 
   const addJobBookmark = () => {
-    return (
-      userId &&
+    if (userId)
       addMutate(
-        { userId, likeType: "jd-bookmarks", elemId: jobData.id },
+        { userId, elemId: jobData.id },
         {
           onSuccess: () => {
-            refetchUserBookmark();
+            queryClient.invalidateQueries(["jobArr"]);
           },
-        }
-      )
-    );
+        } 
+      );
   };
 
   const deleteJobBookmark = () => {
-    return (
-      userId &&
-      deleteMutate(
-        { userId, likeType: "jd-bookmarks", elemId: jobData.id },
-        {
-          onSuccess: () => {
-            refetchUserBookmark();
-          },
-        }
-      )
-    );
+    if (userId) deleteMutate({ userId, likeType: "jd-bookmarks", elemId: jobData.id });
   };
 
   const today = new Date();
